fix: Fail WebSocket extensions handshake when window bits value is not supported
Motivation:

We currently always claim we support window size, but that's not exactly correct.
When using `java.util.zip.Deflater`, the only supported value is 15.
Any other valid value requires falling back to JZlib, which is an optional dependency that might not be in the classpath, leading to a CNFE.

Modification:

Instead of crashing downstream with CNFE, directly crash the extensions handshake if the window bits is not supported because JZlib is missing from the classpath.

Result:

No more cryptic CNFE.